requirements.txt
pandas
geopandas
sqlalchemy
geoalchemy2
psycopg2
python-dotenv